code for 6 channel SB44780
most up to date version for observation SB44780 with the 6 channel run of wsclean. 
no detection of polarization at GRB221009A coordinates at SNRs of 5 for P and 100 for Stokes I.
The GRB is either below the SNRs OR it was channel depolarized due to rotation in the polarization angle.